pyproject.toml
[project]
name = "agent-port"
version = "0.1.0"
description = "uv を使う最小構成の Python CLI サンプル"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "discord-py>=2.4,<3",
]

[dependency-groups]
dev = [
    "pytest>=8.4.0",
]

[tool.pytest.ini_options]
testpaths = ["tests"]
pythonpath = ["."]
